requirements.txt
Django
djangorestframework
drf-yasg
django-ckeditor
django-cors-headers
